+++ src/store/index.ts
@@ -1,0 +1,6 @@
+export type { Direction as EditorDirection, EditorTool } from "./editorStore";
+export { useEditorStore } from "./editorStore";
+export type { Direction as GameDirection } from "./gameStore";
+export { useGameStore } from "./gameStore";
+
+export { useProjectStore } from "./projectStore";